tasks/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloud-sql-python-connector==1.18.2
cloudevents==1.11.1
cryptography==45.0.3
deprecation==2.1.0
dnspython==2.7.0
flask==3.1.1
frozenlist==1.6.0
functions-framework==3.8.3
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-genai==1.17.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
multidict==6.4.4
mypy==1.15.0
mypy-extensions==1.1.0
packaging==25.0
pg8000==1.31.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
rsa==4.9.1
ruff==0.11.11
scramp==1.4.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
types-requests==2.32.0.20250515
typing-extensions==4.13.2
typing-inspection==0.4.1
ulid==1.1
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
websockets==15.0.1
werkzeug==3.1.3
yarl==1.20.0
